Move entity types to @/core
