+++ packages/graph-explorer/src/connector/gremlin/fetchNeighbors/oneHopTemplate.test.ts
@@ -1,6 +1,6 @@
 import { normalizeWithNoSpace as normalize } from "@/utils/testing";
 import oneHopTemplate from "./oneHopTemplate";
-import { VertexId } from "@/@types/entities";
+import { VertexId } from "@/core";
 
 describe("Gremlin > oneHopTemplate", () => {
   it("Should return a template for a simple vertex id", () => {
