+++ packages/graph-explorer/src/connector/gremlin/fetchNeighborsCount/neighborsCountTemplate.test.ts
@@ -1,4 +1,4 @@
-import { VertexId } from "@/@types/entities";
+import { VertexId } from "@/core";
 import neighborsCountTemplate from "./neighborsCountTemplate";
 import { normalizeWithNoSpace as normalize } from "@/utils/testing";
