+++ packages/graph-explorer/src/connector/openCypher/keywordSearch/index.ts
@@ -1,4 +1,4 @@
-import { Vertex } from "@/@types/entities";
+import { Vertex } from "@/core";
 import type {
   ErrorResponse,
   KeywordSearchRequest,
