+++ packages/graph-explorer/src/connector/openCypher/mappers/mapApiEdge.ts
@@ -1,4 +1,4 @@
-import type { Edge, EdgeId, VertexId } from "@/@types/entities";
+import type { Edge, EdgeId, VertexId } from "@/core";
 import type { OCEdge } from "../types";
 
 const mapApiEdge = (
