+++ packages/graph-explorer/src/core/entities.ts
