+++ packages/graph-explorer/src/core/index.ts
@@ -3,3 +3,4 @@ export * from "./ThemeProvider";
 export * from "./featureFlags";
 export * from "./StateProvider";
 export * from "./connector";
+export * from "./entities";
